Change name to shutdownCallbackCore & add comments
