+++ Source/AwsCommonRuntimeKit/auth/imds/CRTIMDSClient.swift
@@ -8,7 +8,7 @@ public class CRTIMDSClient {
     let allocator: Allocator
     public init(options: CRTIMDSClientOptions, allocator: Allocator = defaultAllocator) throws {
         self.allocator = allocator
-        let shutdownOptions = ShutdownCallbackOptions(options.shutdownCallback)?.getIMDSClientShutdownOptions()
+        let shutdownOptions = ShutdownCallbackCore(options.shutdownCallback)?.getIMDSClientShutdownOptions()
                 ?? aws_imds_client_shutdown_options()
         var imdsOptions = aws_imds_client_options(shutdown_options: shutdownOptions,
                                                   bootstrap: options.bootstrap.rawValue,
