+++ Source/AwsCommonRuntimeKit/io/EventLoopGroup.swift
@@ -9,7 +9,7 @@ public final class EventLoopGroup {
     public init(threadCount: UInt16 = 0,
                 allocator: Allocator = defaultAllocator,
                 shutdownCallback: ShutdownCallback? = nil) throws {
-        let shutdownCallbackOptions = ShutdownCallbackOptions(shutdownCallback)
+        let shutdownCallbackOptions = ShutdownCallbackCore(shutdownCallback)
         guard let rawValue: UnsafeMutablePointer<aws_event_loop_group> = withOptionalUnsafePointer(
                 shutdownCallbackOptions?.getCShutdownOptions(), { shutdownOptionsPointer in
             aws_event_loop_group_new_default(allocator.rawValue, threadCount, shutdownOptionsPointer)
